correção de float para double no módulo de serviços

+++ src/modules/servicos/model/Servicos.ts
@@ -29,7 +29,7 @@ function initServico(): void {
         allowNull: false,
       },
       descricao: { type: DataTypes.STRING, allowNull: false },
-      valor: { type: DataTypes.FLOAT, allowNull: false },
+      valor: { type: DataTypes.DOUBLE, allowNull: false },
     },
     { sequelize: connection, tableName: 'servicos', modelName: 'servico' },
   )
